refactor: make PremiumMark more universal
